alembic/requirements.txt
alembic==1.9.2
numpy==1.24.2
pandas==1.5.3
pathlib==1.0.1
psycopg2-binary==2.9.5
SQLAlchemy==2.0.5.post1